- Move ValueMap and AttributeValue to different files
