+++ src/app/submission/submission-shared/model/submission/index.ts
@@ -1,2 +1,4 @@
 export * from './submission.model';
 export * from './submission.validator';
+export * from './submission.model.valuemap';
+export * from './submission.model.attribute-value';
