+++ src/app/submission/submission-shared/model/submission/submission.model.attribute-value.ts
@@ -1,0 +1,5 @@
+export class AttributeValue {
+  constructor(
+    public value: string = ''
+  ) {}
+}
